renamed icons folder for windows
